Refactoring + bug fixes
